Bibie

+++ src/services/user/delete.go
@@ -1,10 +1,16 @@
 package userservice
 
+import "errors"
+
 func (s *userService) Delete(userId string) error {
-	_, findError := s.userRepository.FindById(userId)
+	user, findError := s.userRepository.FindById(userId)
 
 	if findError != nil {
 		return findError
+	}
+
+	if user.Role != "nurse" {
+		return errors.New("THIS USER IS NOT NURSE")
 	}
 
 	err := s.userRepository.Delete(userId)
